+++ libportal/remote.c
@@ -38,7 +38,7 @@ typedef struct {
   guint signal_id;
   GTask *task;
   char *request_path;
-  guint cancelled_id;
+  gulong cancelled_id;
 } CreateCall;
 
 static void
@@ -638,7 +638,7 @@ typedef struct {
   guint signal_id;
   GTask *task;
   char *request_path;
-  guint cancelled_id;
+  gulong cancelled_id;
 } StartCall;
 
 static void
